fix/refactor: alterar css da tabela para scss e ajustar imports
