+++ src/components/table/TableHeader.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { useState } from "react"
 import { TableHeaderProps } from "../../typesTable"
-import themes from '../themes/Table.module.css'
+import themes from '../themes/table.module.scss'
 import editarImg from '../../assets/home/Edit.png'
 import checkImg from '../../assets/checkIcon.svg'
 function TableHeader(props: TableHeaderProps) {
