+++ src/components/table/TableRow.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { RowData, TableRowMethods, TableRowProps } from "../../typesTable"
-import themes from '../themes/Table.module.css'
+import themes from '../themes/table.module.scss'
 import editarImg from '../../assets/home/Edit.png'
 function TableRow(props: TableRowProps) {
     const rowData = props.rowData
