+++ src/components/themes/table.module.scss
